pyproject.toml
[tool.poetry]
name = "dormyboba-core"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "dormyboba_core", from = "." },
]
include = [
    "dormyboba_core/static",
    "dormyboba_core/templates",
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
sqlacodegen = "3.0.0rc3"
psycopg2-binary = "^2.9.9"
pyyaml = "^6.0.1"
grpcio = "^1.60.1"
grpcio-tools = "^1.60.1"
gspread = "^6.0.0"
sqlalchemy = "^2.0.25"
dormyboba-api = {git = "https://github.com/aaletov/dormyboba-api.git", rev = "main", subdirectory = "api/generated/python"}
fastapi = {extras = ["all"], version = "^0.109.2"}
loguru = "^0.7.2"
pyjwt = "^2.8.0"
cryptography = "^42.0.3"
pytest = "^8.0.1"
pytest-cov = "^4.1.0"
dependency-injector = "^4.41.0"


[tool.poetry.group.dev.dependencies]
behave = "^1.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
